pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paircars"
version = "2.0.0"
description = "A package to calibrate MWA solar observation"
readme = { file = "README.md", content-type = "text/markdown" }
license = "MIT"
authors = [
    { name = "Devojyoti Kansabanik", email = "devojyoti96@gmail.com" },
]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]
dependencies = [
    "aiapy==0.10.1",
    "aiobotocore==2.21.1",
    "astroquery==0.4.10",
    "aioftp==0.23.1",
    "aiohappyeyeballs==2.4.3",
    "aiohttp==3.10.10",
    "aioitertools==0.12.0",
    "aiosignal==1.3.1",
    "antlr4-python3-runtime==4.9.3",
    "appdirs==1.4.4",
    "asciitree==0.3.3",
    "astLib==0.11.10",
    "astro-kittens==1.4.6",
    "astro-tigger-lsm==1.7.3",
    "astropy==6.1.4",
    "astropy-iers-data==0.2024.10.28.0.34.7",
    "astropy_healpix==1.1.2",
    "asttokens==2.4.1",
    "async-timeout==4.0.3",
    "attrs==24.2.0",
    "beautifulsoup4==4.13.4",
    "black==24.4.2",
    "bokeh==3.6.3",
    "botocore==1.37.1",
    "bs4==0.0.2",
    "casaconfig==1.0.0",
    "casadata==2024.5.22",
    "casatasks==6.6.0.20",
    "casatools==6.6.0.20",
    "certifi==2024.7.4",
    "charset-normalizer==3.4.1",
    "cli-ui==0.18.0",
    "click==8.1.7",
    "cloudpickle==3.0.0",
    "codex-africanus==0.3.7",
    "colorama==0.4.6",
    "Columnar==1.4.1",
    "configparser==7.1.0",
    "contourpy==1.2.1",
    "cycler==0.12.1",
    "dask==2024.7.1",
    "dask-jobqueue==0.9.0",
    "dask-mpi==2022.4.0",
    "dask-ms==0.2.21",
    "decorator==5.1.1",
    "dill==0.3.9",
    "distributed==2024.7.1",
    "docopt==0.6.2",
    "donfig==0.8.1.post1",
    "drms==0.8.0",
    "exceptiongroup==1.2.2",
    "executing==2.0.1",
    "fasteners==0.19",
    "fonttools==4.53.1",
    "frozenlist==1.5.0",
    "fsspec==2025.2.0",
    "future==1.0.0",
    "git-filter-repo==2.47.0",
    "idna==3.10",
    "importlib_metadata==8.2.0",
    "iniconfig==2.0.0",
    "ipython==8.26.0",
    "isodate==0.7.2",
    "jedi==0.19.1",
    "Jinja2==3.1.6",
    "jmespath==1.0.1",
    "joblib==1.4.2",
    "julian==0.14",
    "kiwisolver==1.4.5",
    "llvmlite==0.43.0",
    "locket==1.0.0",
    "loguru==0.7.2",
    "lxml==5.4.0",
    "markdown-it-py==3.0.0",
    "MarkupSafe==3.0.2",
    "matplotlib==3.5.2",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "mpl_animators==1.2.1",
    "msgpack==1.1.0",
    "multidict==6.1.0",
    "munch==2.5.0",
    "mypy-extensions==1.0.0",
    "numba==0.60.0",
    "numcodecs==0.13.1",
    "numpy==1.26.4",
    "omegaconf==2.3.0",
    "packaging==24.1",
    "pandas==2.2.3",
    "parfive==2.1.0",
    "parso==0.8.4",
    "partd==1.4.2",
    "pathspec==0.12.1",
    "pexpect==4.9.0",
    "pillow==10.4.0",
    "platformdirs==4.2.2",
    "pluggy==1.5.0",
    "prompt_toolkit==3.0.47",
    "propcache==0.2.0",
    "psutil==5.9.8",
    "ptyprocess==0.7.0",
    "pure_eval==0.2.3",
    "pydantic>=2.10.1",
    "pyerfa==2.0.1.4",
    "Pygments==2.18.0",
    "pyparsing==3.1.2",
    "python-casacore==3.6.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.2",
    "PyWavelets==1.8.0",
    "PyYAML==6.0.1",
    "regions==0.9",
    "reproject==0.14.1",
    "requests==2.32.3",
    "requests-file==2.1.0",
    "requests-toolbelt==1.0.0",
    "rich==13.9.4",
    "ruamel.yaml==0.18.6",
    "ruamel.yaml.clib==0.2.12",
    "s3fs==2025.2.0",
    "schema==0.7.7",
    "scipy==1.15.2",
    "six==1.16.0",
    "sortedcontainers==2.4.0",
    "soupsieve==2.6",
    "stack-data==0.6.3",
    "sunpy[all]==6.0.5",
    "tabulate==0.9.0",
    "tblib==3.0.0",
    "tbump==6.11.0",
    "threadpoolctl==3.5.0",
    "tomli==2.0.1",
    "tomlkit==0.11.8",
    "toolz==0.12.1",
    "tornado==6.4.2",
    "tqdm==4.66.6",
    "traitlets==5.14.3",
    "typeguard==4.4.2",
    "typing_extensions==4.12.2",
    "tzdata==2024.2",
    "udocker==1.3.17",
    "Unidecode==1.3.8",
    "urllib3==2.3.0",
    "wcwidth==0.2.13",
    "wrapt==1.17.2",
    "xarray==2025.1.2",
    "xyzservices==2025.1.0",
    "yarl==1.17.1",
    "zarr==2.18.3",
    "zeep==4.3.1",
    "zict==3.0.0",
    "zipp==3.19.2",
    "h5py==3.13.0",
    "moviepy==2.1.2",
    "h5netcdf==1.6.1",
    "gunicorn==23.0.0",
    "flask==3.1.1",
    "flask-session==0.8.0",
    "watchdog==6.0.0",
    "PyQT5==5.15.11",
    "prefect==3.4.9",
    "prefect-dask==0.3.6",
    "python-dotenv==1.1.1",
]
urls = { "Homepage" = "https://github.com/devojyoti96/P-AIRCARS" }

[project.optional-dependencies]
dev = [
    "sphinx>=7.2.6",
    "furo>=2024.5.6",                      
    "sphinx_rtd_theme>=2.0.0",             
    "sphinx-autodoc-typehints>=2.0.0",
    "sphinx-copybutton>=0.5.2",
    "sphinxcontrib-programoutput>=0.17",
    "sphinx-togglebutton>=0.3.2",
    "myst-parser>=2.0.0",
    "sphinxcontrib-apidoc>=0.4.0",
    "sphinx-automodapi>=0.15.0",
    "numpydoc>=1.6.0",
    "sphinxcontrib-mermaid>=0.9.2",
    "graphviz>=0.20.3",                    
    "sphinx-autobuild>=2021.3.14",
    "pytest==8.3.3",
    "pytest-mock>=3.14.0",
]


# Console scripts entry points
[project.scripts]
run-mwa-import-model = "paircars.pipeline.import_model:cli"
init-paircars-setup = "paircars.pipeline.init_data:cli"
init-paircars-prefect="paircars.pipeline.prefect_server:cli"
kill-paircars-job = "paircars.pipeline.kill_job:kill_paircarsjob"

[tool.setuptools]
packages = [
    "paircars",
    "paircars.utils",
    "paircars.pipeline",
    "paircars.data"
]
include-package-data = true

[tool.setuptools.package-data]
"paircars.data" = ["sendmail.so", "hyperdrive"]



